rename default docmapper to docmapper
